pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-mssql-backend = "^2.8.1"
psycopg2-binary = "^2.9.9"
mysqlclient = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
